Moved dedicated platforms views to ItemContentDetails.
